% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Bisep_Plot.R
\name{BicepPlot}
\alias{BicepPlot}
\title{Function to create Bi-Separation plot for comparing two models}
\usage{
BicepPlot(f1, f2, y, labels, bestN = 10, label_spacing = 3,
  right_lab_adjust = 0.02, bottom_lab_adjust = 0.03, right_margin = 5,
  bottom_margin = 5, top_margin = 1, transp_adjust = 1,
  m1title = "Model 1", m2title = "Model 2", hlines = TRUE,
  vlines = TRUE, rare = FALSE)
}
\arguments{
\item{f1}{Vector of forecasts from model 1}

\item{f2}{Vector of forecasts from model 2}

\item{y}{Vector of actual outcomes}

\item{labels}{Vector of labels per observation}

\item{bestN}{max num. of "most improved" obs you want labeled per model}

\item{label_spacing}{= distance between labels}

\item{right_lab_adjust}{how far from the end of line do you want label? Usually small; default is 0.02}

\item{bottom_lab_adjust}{how far from end of line do you want label?Usually small; default is 0.03}

\item{right_margin}{how much whitespace needed on right side of plot? For if labels don't fit; default is 5}

\item{bottom_margin}{how much whitespace needed on bottom of plot? For if labels don't fit; default is 5}

\item{top_margin}{how much whitespace needed on top of plot? For if labels don't fit; default is 1}

\item{transp_adjust}{float from [0,10] to adjust opacity of non-labeled points. Default is 1. Generally on the extremes (1 vs 10). 0 will render them extremely light}

\item{hlines}{True/False whether you want faded lines horizontally pointing from M2 separation plot to most improved observations. Default is TRUE.}

\item{vlines}{True/False whether you want faded lines vertically pointing from M1 separation plot to most improved observations. Default is TRUE}

\item{rare}{True/False whether you have an excessive number of zeros. Adjusts coloring for visibility. Default is FALSE}
}
\value{
Object to be plotted with "grid.draw()' function
}
\description{
Creates an object to be rendered with "grid.draw()" command
}

