% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Diag_Plot.R
\name{DiagPlot}
\alias{DiagPlot}
\title{Function to create diagnostic plot for a single model}
\usage{
DiagPlot(f, y, labels, worstN = 10, size_adjust = 0, right_margin = 7,
  top_margin = 1.5, label_spacing = 10, lab_adjust = 0.5,
  text_size = 10, title = "Model Diagnostic Plot")
}
\arguments{
\item{f}{N by 1 vector of forecasts f_{i} in (0,1)}

\item{y}{N by 1 vector of forecasts y_{i} is 0 or 1}

\item{labels}{N by 1 vector of textual labels (strings)}

\item{worstN}{How many values to label, given y={0,1}}

\item{size_adjust}{adjusts distribution of space between density plot and the main plot. Usually values from [0,.7], but you can also go into negative numbers where you need the margx plot (density plot on x axis) to get larger relative to main plot. >0.6 starts getting ugly. Default is 0.5.}

\item{right_margin}{adjusts the right margin. Tends to be between [5,15]. Default value is 7.}

\item{top_margin}{adjusts the top margin of PLOT (not including title). Useful if labels are off the upper end of page. Likely between [1,5]. Default value is 1.5.}

\item{label_spacing}{Spacing between labels}

\item{lab_adjust}{shift label right or left vis-a-vis the line}

\item{text_size}{font size}
}
\value{
Object to be plotted with "grid.draw() function
}
\description{
Creates an object to be rendered with "grid.draw()" command
}

