% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_coxph.R
\name{extract_coxph}
\alias{extract_coxph}
\title{function to extract relevant components from a coxph model}
\usage{
extract_coxph(lm, alpha = 0.05, output = FALSE)
}
\description{
function to extract relevant components from a coxph model
}
