% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lm.R
\name{extract_lm}
\alias{extract_lm}
\title{Extract Relevant lm Components}
\usage{
extract_lm(lm, alpha = 0.05, output = TRUE)
}
\arguments{
\item{lm}{linear model}

\item{alpha}{user-defined alpha; define the threshold for significance. Default is 0.05.}

\item{output}{controls whether the extracted components are published in the console, default is 'TRUE'.}
}
\value{
dataframe of relevant lm components
}
\description{
extract relevant components from a linear model
}
\details{
I want the user to be able to define the alpha value, so some resulting values in the general
summary(lm) would need to be made dependent on that change from 0.05).  The output will be used
downstream to supply model information in a display/table for comparison against other similar/nested models.
}
\examples{
#fit linear model
 lmod<- lm(gamble ~ sex + status + income + verbal + sex:income, data = teengamb)
#extract components, supplying 'output=TRUE' to print output
 extract_lm(lmod)

}
