% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablecomp.R
\name{tablecomp}
\alias{tablecomp}
\title{Generate Model Comparison Table}
\usage{
tablecomp(
  ...,
  alpha_ = 0.05,
  modeltype = c("lm", "coxph"),
  comparison_value = c("coefs", "stderrs", "t_vals", "p_vals", "stars", "lower_confints",
    "higher_confints", "rsq", "adj.rsq", "aic", "alpha")
)
}
\arguments{
\item{alpha_}{user-defined alpha; define the threshold for significance. Default is 0.05.}

\item{modeltype}{'lm' for linear model, 'coxph' for cox proportional hazards model.}

\item{comparison_value}{model components to be compared in the output table. Can specify one, multiple, or all.}

\item{model_extract}{model(s) to display components for. can be just one, or as many as desired for side-by-side comparison of values}
}
\value{
table of relevant model components for a quick side-by-side comparison between models
}
\description{
Generate a single table for easy comparison between desired models, where each model occupies its own column/ set of columns.
}
\details{
User can define which model components are of interest for a quick comparison, for example, model coefficient estimats ('coefs') and associated p values ('p_vals'). All components extracted
using 'extract_lm()' are available for comparison in this method. One quirk here is that models are displayed left to right in the order they are supplied in the funtion. A later version would fix the
lack of model labels, but for now this is the method to identify the models being compared--by order. It is also advised to supply the largest model first (especially if comparing nested models), as it will
ensure all predictor names will be displayed successfully. A later version would fix this quirk as well.
}
\examples{
#supply linear model(s) for output comparison
 lmod1<- lm(gamble ~ sex + status + income + verbal + sex:status + sex:income + sex:verbal, data = faraway_teengamb)
 lmod2<- lm(gamble ~ sex + status + income + verbal + sex:income, data = faraway_teengamb)
 lmod3<- lm(gamble ~ sex + status + income + verbal, data = faraway_teengamb)
#determine what comparison_value(s) are important for the table, or user can do one comparison value per table to make viewing even easier. Then create desired table(s) using 'comptable()'.
tablecomp(lmod1)
tablecomp(lmod1, lmod2, comparison_value= 'coefs')
tablecomp(lmod1, lmod2, lmod3, comparison_value= c('coefs', 'p_vals', 'stars'))


}
