% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_lm.R
\name{extract_lm}
\alias{extract_lm}
\title{Extract Relevant lm Components}
\usage{
extract_lm(lm, alpha = 0.05, output = FALSE)
}
\arguments{
\item{lm}{linear model}

\item{alpha}{user-defined alpha; define the threshold for significance}

\item{output}{controls whether the extracted components are published, default is that the output is not published, but can be set to 'TRUE' to publish results}
}
\value{
list of relevant lm components, if output= TRUE is supplied by user, otherwise, output is not published
}
\description{
extract relevant components from a linear model
}
\details{
I want the user to be able to define the alpha value, so some resulting values in the general
summary(lm) would need to be made dependent on that change from 0.05).  The output will be used
downstream to supply model information in a display/table for comparison against other similar/nested models.
}
