% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablestack.R
\name{tablestack}
\alias{tablestack}
\title{Generate a stack of model output tables for quick comparison.}
\usage{
tablestack(..., alpha_ = 0.05, modeltype = "lm")
}
\arguments{
\item{...}{model(s) to display components for. can be just one, or as many as desired for a comparison of values}

\item{alpha_}{user-defined alpha; define the threshold for significance. Default is 0.05.}

\item{modeltype}{'lm' for linear model. Future versions will include other model types (glm, coxph, for example)}
}
\value{
table(s) of relevant model components for a quick comparison. They are displayed in the order they are entered as input, with the top being the first model specified, and so on.
}
\description{
stack model information tables on top of each other to allow for a quick comparison between values.
}
\details{
function to generate the output of multiple models (of the same class) stacked on top of each other to make comparison of values easier. Or, only one model can be specified and the output
will still list the details for the singular model in an efficient way. This makes one large improvement over 'summary(lm)', as the user can define an alpha value beyond just a set 0.05.
}
\examples{
#fit linear models
data(faraway_teengamb)
lmod1<-lm(gamble~sex+status+income+verbal+sex:status+sex:income+sex:verbal,data=faraway_teengamb)
lmod2<-lm(gamble~sex+status+income+verbal+sex:income, data = faraway_teengamb)
#use 'tablestack()' to compare outputs displayed; can choose a user-defined alpha
#if default 0.05 is not the desired level.
 tablestack(lmod1, lmod2, alpha_= 0.1)

}
